requirements.txt
docopt==0.6.2
num2words==0.5.12
Pillow==9.5.0
